+++ app/src/main/java/io/oldering/tvfoot/red/matches/MatchesActivity.java
@@ -8,6 +8,9 @@ import io.oldering.tvfoot.red.R;
 import io.oldering.tvfoot.red.databinding.ActivityMatchesBinding;
 import io.oldering.tvfoot.red.flowcontroller.FlowController;
 import io.oldering.tvfoot.red.matches.displayable.MatchRowDisplayable;
+import io.oldering.tvfoot.red.matches.state.MatchesIntent;
+import io.oldering.tvfoot.red.matches.state.MatchesStateManager;
+import io.oldering.tvfoot.red.matches.state.MatchesViewState;
 import io.reactivex.Observable;
 import io.reactivex.disposables.CompositeDisposable;
 import javax.annotation.Nullable;
