+++ app/src/main/java/io/oldering/tvfoot/red/matches/MatchesAdapter.java
@@ -16,6 +16,7 @@ import io.oldering.tvfoot.red.matches.displayable.LoadingRowDisplayable;
 import io.oldering.tvfoot.red.matches.displayable.MatchRowDisplayable;
 import io.oldering.tvfoot.red.matches.displayable.MatchesItemDisplayable;
 import io.oldering.tvfoot.red.matches.displayable.MatchesItemDisplayableDiffUtilCallback;
+import io.oldering.tvfoot.red.matches.state.MatchesIntent;
 import io.reactivex.Observable;
 import io.reactivex.subjects.PublishSubject;
 import java.util.Collections;
