+++ app/src/test/java/io/oldering/tvfoot/red/di/component/TestComponent.java
@@ -5,7 +5,7 @@ import io.oldering.tvfoot.red.di.module.NetworkModule;
 import io.oldering.tvfoot.red.di.module.ServiceMockModule;
 import io.oldering.tvfoot.red.di.module.TestSchedulerModule;
 import io.oldering.tvfoot.red.match.MatchInteractor;
-import io.oldering.tvfoot.red.matches.MatchesService;
+import io.oldering.tvfoot.red.matches.state.MatchesService;
 import io.oldering.tvfoot.red.util.Fixture;
 import io.oldering.tvfoot.red.util.schedulers.BaseSchedulerProvider;
 import javax.inject.Singleton;
